Packages refactoring. Added class Utils for reading input data.
