+++ src/main/java/com/advent/of/code/day1/Measurements.java
@@ -1,4 +1,4 @@
-package day2;
+package com.advent.of.code.day1;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
@@ -7,14 +7,14 @@ import java.nio.file.FileSystems;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CoursePlan {
-    public static List<String> generate(String filename) {
-        List<String> resultList = new ArrayList<>();
-        final String path = FileSystems.getDefault().getPath("").toAbsolutePath().toString().concat("\\src\\main\\java\\day2\\" + filename);
+public class Measurements {
+    public static List<Integer> generate(String filename) {
+        List<Integer> resultList = new ArrayList<>();
+        final String path = FileSystems.getDefault().getPath("").toAbsolutePath().toString().concat("\\src\\main\\java\\day1\\" + filename);
         try (BufferedReader bufferedReader = new BufferedReader(new FileReader(path))) {
             String currentLine;
             while ((currentLine = bufferedReader.readLine()) != null) {
-                resultList.add(currentLine);
+                resultList.add(Integer.parseInt(currentLine));
             }
         } catch (IOException e) {
             e.printStackTrace();
