+++ src/main/java/com/advent/of/code/day1/input.txt
