+++ src/main/java/com/advent/of/code/day2/input.txt
