+++ src/main/java/com/advent/of/code/day3/Report.java
@@ -1,4 +1,4 @@
-package day3;
+package com.advent.of.code.day3;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
