+++ src/main/java/com/advent/of/code/day3/Solution.java
@@ -1,4 +1,4 @@
-package day3;
+package com.advent.of.code.day3;
 
 public class Solution {
     public static void main(String[] args) {
