+++ src/main/java/com/advent/of/code/day3/input.txt
